fix: bind to a specific local IPv6 address during pair verify
This will prevent future traffic from picking the "best" address to
communicate with an accessory. Current products tie sessions to the
original IPv6 address and return a CoAP error code 4.04 when a new
address attempts to continue the session.

It is likely that this could explode in fun, new ways: the DHCPv6 server
returning a new address, the host selecting a new random SLAAC address,
etc.