Implement async_reachable to see if a device is reachable on the network
For zeroconf based devices we check if the A/AAAA records are still
published. This means we will know if the device is offline within
120s since A/AAAA records typically have a TTL of 120s

For BLE based devices we will check if the BLE device is in range and
seen by the adapter. This will vary by BLE adapter, but it will likely
mean we will know within a few minutes as well